assignment 2

+++ Quiz1/Problem3/P3_1.c
@@ -3,7 +3,7 @@ int main(){
     char x[64];
     long int *a=(long int*)&x;
     long int *b=(long int*)&(x[64]);
-    int val=1000;
+    long int val=1000;
     for(; a!=b;a++){
         *a=val;
         printf("%ld\n",*a);
